requirements.txt
# https://github.com/GregDMeyer/IT8951
Pillow
tensorflow
jax
objax
tqdm
numpy
joblib